+++ LibraryDX12/Window/SwapChain.h
@@ -74,5 +74,5 @@ private:
     // Numerical data
     uint32_t width;
     uint32_t height;
-    size_t currentBackBufferIndex;
+    UINT currentBackBufferIndex;
 };
